requirements.txt
absl-py==2.1.0
advocate==1.0.0
alembic==1.13.1
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
APScheduler==3.10.4
arabic-reshaper==3.0.0
argcomplete==3.4.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asn1crypto==1.5.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==23.2.0
Babel==2.15.0
bcrypt==4.1.3
beautifulsoup4==4.12.3
bleach==6.1.0
blinker==1.8.2
CacheControl==0.14.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.2
click==8.1.7
cloudmersive_barcode_api_client==3.1.0
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.7
cssselect==1.2.0
cssselect2==0.7.0
curses-util==0.0.25
cycler==0.12.1
Cython==3.0.10
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
demoji==1.1.0
Deprecated==1.2.14
distlib==0.3.8
Django==5.0.6
executing==2.0.1
fastjsonschema==2.20.0
feature-engine==1.8.0
filelock==3.14.0
firebase-admin==6.5.0
flasgger==0.9.7.1
Flask==2.3.3
flask-babel==4.0.0
Flask-Cors==4.0.1
Flask-Limiter==3.5.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Moment==0.11.0
Flask-MySQLdb==2.0.0
Flask-Principal==0.4.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
fsspec==2024.6.0
gast==0.5.4
gcloud==0.18.3
google-ai-generativelanguage==0.6.4
google-api-core==2.19.0
google-api-python-client==2.48.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-generativeai==0.6.0
Google-Images-Search==1.4.6
google-pasta==0.2.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.1
grpcio==1.64.0
grpcio-status==1.62.2
gunicorn==22.0.0
h11==0.14.0
h5py==3.11.0
html5lib==1.1
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
imbalanced-learn==0.12.3
importlib_resources==6.4.0
imutils==0.5.4
install==1.3.5
ipykernel==6.29.4
ipython==8.25.0
ipywidgets==8.1.3
iso-639==0.4.5
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.1
jupyter_server_terminals==0.5.3
jupyterlab==4.2.3
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.11
jwcrypto==1.5.6
keras==3.3.3
kiwisolver==1.4.5
libclang==18.1.1
limits==3.12.0
lxml==5.1.1
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.1
matplotlib-inline==0.1.7
maturin==1.7.0
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.3.2
mpmath==1.3.0
msgpack==1.0.8
mysql-connector-python==8.4.0
mysqlclient==2.2.4
namex==0.0.8
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
ndg-httpsclient==0.5.1
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.3
nltk==3.8.1
notebook==7.2.1
notebook_shim==0.2.4
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
ordered-set==4.1.0
oscrypto==1.3.0
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.0
pandas==2.2.2
pandocfilters==1.5.1
parsel==1.9.1
parso==0.8.4
patsy==0.5.6
pexpect==4.9.0
pillow==10.3.0
pipx==1.6.0
platformdirs==4.2.2
plotly==5.22.0
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.4
pydantic_core==2.18.4
pydotplus==2.0.2
pyfiglet==0.8.post1
Pygments==2.18.0
pyHanko==0.25.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
PyMySQL==1.1.0
pyOpenSSL==24.1.0
pyparsing==3.1.2
pypdf==4.0.2
pypng==0.20220715.0
Pyrebase4==4.8.0
PySocks==1.7.1
python-bidi==0.4.2
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-jwt==4.1.0
python-resize-image==1.1.20
pytz==2021.1
pyvi==0.1.1
PyYAML==6.0.1
pyzbar==0.1.9
pyzmq==26.0.3
qrcode==7.4.2
qrdet==2.4
qreader==3.13
qtconsole==5.5.2
QtPy==2.4.1
quadrilateral-fitter==1.8
referencing==0.35.1
regex==2023.12.25
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==2.0.0
requests-toolbelt==0.10.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
scikit-learn==1.5.1
scipy==1.14.0
seaborn==0.13.2
selenium==4.21.0
Send2Trash==1.8.3
setuptools==70.2.0
shapely==2.0.4
six==1.16.0
sklearn-crfsuite==0.5.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
stack-data==0.6.3
statsmodels==0.14.2
svglib==1.5.1
sympy==1.12.1
tabulate==0.9.0
tenacity==8.5.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
termcolor==1.1.0
terminado==0.18.1
threadpoolctl==3.5.0
tinycss2==1.3.0
torch==2.3.1
torchvision==0.18.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
trio==0.25.1
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ultralytics==8.2.31
ultralytics-thop==0.2.8
Unidecode==1.3.8
uri-template==1.3.0
uritemplate==4.1.1
uritools==4.0.2
urllib3==1.26.18
userpath==1.9.2
virtualenv==20.26.2
w3lib==2.2.1
Wand==0.6.13
wcwidth==0.2.13
webcolors==24.6.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.0.3
wheel==0.43.0
widgetsnbextension==4.0.11
wordcloud==1.9.3
wrapt==1.16.0
wsproto==1.2.0
WTForms==3.1.2
xgboost==2.1.0
xhtml2pdf==0.2.15
yellowbrick==1.5
youtube-search-python==1.6.6
